requirements.txt
sanic~=20.9.0
sanic-mongodb-extension
sanic-jinja2
sanic-openapi

umongo[motor]~=2.3.0
motor~=2.3.0
dnspython